# docs/swagger.yaml
basePath: /
definitions:
...
  db.UpdateUser:
    properties:
      is_admin:
        type: integer
      login:
        type: string
    required:
    - login
    type: object
  db.User:
    properties:
      id:
        type: integer
      is_admin:
        type: integer
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
host: localhost:8888
info:
  contact: {}
  description: RESTful API server for MIFARE transactions authorization
  title: MIFARE API
  version: "0.1"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Возвращает список всех пользователей (доступно всем авторизованным)
      produces:
      - application/json
      responses:
        "200":
          description: Список пользователей в поле data
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение всех пользователей
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Создает нового пользователя (требуется роль администратора)
      parameters:
      - description: Данные пользователя (login, password, role)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/db.User'
      produces:
      - application/json
      responses:
        "200":
          description: ID созданного пользователя
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Неверные параметры
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Недостаточно прав (требуется администратор)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создание пользователя
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет пользователя по ID (требуется роль администратора)
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Успешно удалено
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Недостаточно прав (требуется администратор)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удаление пользователя
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Возвращает данные пользователя по идентификатору (доступно всем
        авторизованным)
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные пользователя
          schema:
            $ref: '#/definitions/db.User'
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение пользователя по ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Обновляет данные пользователя (требуется роль администратора)
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/db.UpdateUser'
      produces:
      - application/json
      responses:
        "204":
          description: Успешно обновлено
        "400":
          description: Неверные параметры или ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Недостаточно прав (требуется администратор)
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновление пользователя
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"